refactor: modularize work order app
