+++ src/components/DateField.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+export default function DateField({ label, value, onChange }: { label: string; value: string; onChange: (v: string) => void }) {
+  return (
+    <div>
+      <label className="block text-sm text-slate-300 mb-1">{label}</label>
+      <input
+        type="date"
+        value={value}
+        onChange={(e) => onChange(e.target.value)}
+        className="px-3 py-2 rounded-xl bg-slate-950 border border-slate-700 w-full"
+      />
+    </div>
+  );
+}
